fix: add `configs` directory and sample workspace
We add the `:/configs` directory which contains things like application
default themes, shortcuts, and workspaces (needed to load a bare
application).

The Classic.mws workspace is the default one from Audacity. We need to
adapt it later.

This commit also removes the audioplugins and cloud modules since we
don't need them and one doesn't even work.